requirements.txt
# Core dependencies for Vision PyTorch Lightning Framework

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0
torchinfo>=1.8.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Experiment Tracking
mlflow>=2.5.0
tensorboard>=2.13.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Progress and Logging
rich>=13.4.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# Optional accelerators
# For GPU support (already included with torch)
# For TPU support
# torch-xla>=2.0.0  # Uncomment for TPU support

# Additional utilities
pyyaml>=6.0
python-dotenv>=1.0.0
wandb>=0.15.0  # Optional: for Weights & Biases integration